feat(castor): use `local` file for local methods
